deal with delaware river bridges
calcualte bridge buffer in data setup and use it as a filter for which OD pairs to calculate shortest paths for in 4 (island analysis). this prevents millions of long paths going out of their way to get to a bridge to get to their destination in a different state. they caon only do so if they are close to the bridge.